Move playback controls to separate component
